Rating fot workers is working

+++ app/src/main/java/com/example/user/jobche/Model/Application.kt
@@ -4,7 +4,7 @@ import com.example.user.jobche.Task
 
 data class Applications(val applications: ArrayList<Application>)
 
-data class Application(val id: Int,
+data class Application(val id: Long,
                        val accepted: Boolean,
                        val applicant: UserProfile,
                        val task: Task
