+++ app/src/main/java/com/example/user/jobche/OpenedTaskViewModel.kt
@@ -9,7 +9,7 @@ import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class OpenedTaskViewModel(val task: Task, private val email: String, private val password: String, private val userId: Int) {
+class OpenedTaskViewModel(val task: Task, private val email: String, private val password: String, private val userId: Long) {
 
     val date: String = (task.dateTime).substring(8, 10) + "." + (task.dateTime).substring(5, 7)
 //    val date: String = task.dateTime!!.dayOfMonth.toString() + "." + task.dateTime!!.monthOfYear.toString()
