+++ app/src/main/java/com/example/user/jobche/UI/Fragments/TaskAppliedFragment.kt
@@ -97,8 +97,8 @@ class TaskAppliedFragment : Fragment(), AppliersRecyclerViewAdapter.OnApplierCli
     override fun onClick(position: Int) {
         bundle = Bundle()
         newFragment = ApplierProfileFragment()
-        bundle.putInt("ApplicationId", taskAppliersViewModel.applications[position].id)
-        bundle.putInt("ApplierId", appliers[position].id)
+        bundle.putLong("ApplicationId", taskAppliersViewModel.applications[position].id)
+        bundle.putLong("ApplierId", appliers[position].id)
         bundle.putString("Name", appliers[position].firstName)
         newFragment.arguments = bundle
         activity!!.supportFragmentManager.beginTransaction().replace(
